Handle failure to read EFI variables more gracefully
Up to now, we just called fatal() and had the program abort.
Instead, return an error and allow the upper layers to process the
entire event log and *then* exit with status 1.

In testing ALP as a VM, I'm regularly seeing a behavior where the
BootOrder references Boot0003, but that entry does not exist
(any longer). On the next reboot, everything is aligned again...

Signed-off-by: Olaf Kirch <okir@suse.com>