Add new option, --rsa-bits, to set RSA key size
This commit adds the new option, --rsa-bits, to set the RSA key size of
SRK and the signing key. Supported key sizes are 2048, 3072, and 4096
bits.

In OS runtime, RSA 2048 and RSA 4096 are commonly used. However, due to
the size limit of Non-volatile memory, RSA 4096 is not widely supported
in the TPM 2.0 chips till now. Per "TCG PC Client Platform TPM Profile
Specification for TPM 2.0" ver. 1.05, RSA 2048 and RSA 3072 are
mandatory for the newer TPM 2.0 chips. Thus, RSA 3072 is added to the
supported list, and it's recommended to use RSA 3072 if RSA 4096 is not
available.

Signed-off-by: Gary Lin <glin@suse.com>